# Specialize GPUArrays.global_index() to improve broadcast performance