--- a/src/gpuarrays.jl
+++ b/src/gpuarrays.jl
@@ -29,10 +29,11 @@ end
 ## on-device
 
 # indexing
-GPUArrays.blockidx(ctx::mtlKernelContext)  = Metal.threadgroup_position_in_grid_1d()
-GPUArrays.blockdim(ctx::mtlKernelContext)  = Metal.threads_per_threadgroup_1d()
-GPUArrays.threadidx(ctx::mtlKernelContext) = Metal.thread_position_in_threadgroup_1d()
-GPUArrays.griddim(ctx::mtlKernelContext)   = Metal.threadgroups_per_grid_1d()
+GPUArrays.blockidx(ctx::mtlKernelContext)     = Metal.threadgroup_position_in_grid_1d()
+GPUArrays.blockdim(ctx::mtlKernelContext)     = Metal.threads_per_threadgroup_1d()
+GPUArrays.threadidx(ctx::mtlKernelContext)    = Metal.thread_position_in_threadgroup_1d()
+GPUArrays.griddim(ctx::mtlKernelContext)      = Metal.threadgroups_per_grid_1d()
+GPUArrays.global_index(ctx::mtlKernelContext) = Metal.thread_position_in_grid_1d()
 
 # math