@maxwindiff 2023-01-27
This doesn't seem to help with https://github.com/JuliaGPU/Metal.jl/issues/41 however.